requirements.txt
python-dotenv
google-generativeai
pillow
streamlit
tensorflow
numpy
